requirements.txt
flask
pandas 
numpy
pyPDF2
python-docx
anthropic